+++ src/sympc/algorithms/__init__.py
@@ -1,3 +1,3 @@
 """Applications of primitive MPC operations."""
 
-from . import sort
+from . import algorithms
