+++ src/sympc/algorithms/algorithms.py
